refactor: consolidate Doc2Vec components into swarmauri_vectorstore_doc2vec package
