+++ pkgs/standards/swarmauri_standard/pyproject.toml
@@ -88,7 +88,7 @@ full = [
     #"transformers",
     #"torch",
     #"keras", "tf-keras",
-    "matplotlib"
+    #"matplotlib"
 ]
 
 [tool.poetry.group.dev.dependencies]
