+++ pkgs/standards/swarmauri_vectorstore_doc2vec/README.md
