+++ pkgs/standards/swarmauri_vectorstore_doc2vec/pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "swarmauri_doc2vec_vectorstore"
+name = "swarmauri_vectorstore_doc2vec"
 version = "0.6.0.dev1"
 description = "A Doc2Vec based Vector Store and Doc2Vec Based Embedding Model."
 authors = ["Jacob Stewart <jacob@swarmauri.com>"]
@@ -52,7 +52,7 @@ log_cli_date_format = "%Y-%m-%d %H:%M:%S"
 asyncio_default_fixture_loop_scope = "function"
 
 [tool.poetry.plugins."swarmauri.vector_stores"]
-Doc2VecVectorStore = "swarmauri_doc2vec_vectorstore:Doc2VecVectorStore"
+Doc2VecVectorStore = "swarmauri_vectorstore_doc2vec:Doc2VecVectorStore"
 
 [tool.poetry.plugins."swarmauri.embeddings"]
-Doc2VecEmbedding = "swarmauri_doc2vec_vectorstore:Doc2VecEmbedding"
+Doc2VecEmbedding = "swarmauri_vectorstore_doc2vec:Doc2VecEmbedding"
