+++ pkgs/standards/swarmauri_vectorstore_doc2vec/swarmauri_vectorstore_doc2vec/Doc2VecEmbedding.py
