+++ pkgs/standards/swarmauri_vectorstore_doc2vec/tests/unit/Doc2VecEmbedding_unit_test.py
@@ -1,5 +1,5 @@
 import pytest
-from swarmauri_doc2vec_vectorstore.Doc2VecEmbedding import Doc2VecEmbedding
+from swarmauri_vectorstore_doc2vec.Doc2VecEmbedding import Doc2VecEmbedding
 
 
 @pytest.mark.unit
